Created cGearPath and cPlaceGear. Added variables.
cGearPath does a bunch of math to find the path. cPlaceGear actually
moves the robot along this path. I deleted cGoToHook because it was made
redundant. I also added a bunch of variables to AutoBase and RobotMap.
Also fixed a minor thing in cTargetingComputer. There are many commented
out things, which are commented out because they are not yet relevant.